+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Fix ios it should support tls 1.3
 - Fix lpc it should support generic encryption algorithm, tested with: rsa2048, rsa4096, ecdsa256, ecdsa384
 - Fix android it should not play ringtone if there is a connected call (issue 1177)
+- Fix it should work well with special character in park number (issue 1195)
 
 #### 2.17.0